pii/projet 3/pii3/openai-rfp-response-analyzer-main/requirements.txt
django==5.0.3
python-dotenv==1.0.1
uvicorn==0.27.1
httpx==0.24.1
asgiref==3.7.2
django-cors-headers==4.3.1
spacy==3.7.2
python-stdnum==1.20
presidio-analyzer==2.2.32
presidio-anonymizer==2.2.32
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0-py3-none-any.whl
llama-parse==0.5.1
langchain>=0.0.350
langchain-openai>=0.0.5
openai>=1.12.0
pydantic>=2.0.0
transformers>=4.38.2
langsmith>=0.1.0
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
